State machine
The inventory system and console no longer work on a closed circuit type system. It's a lot cleaner because of the state machine approach. Tomorrow I'll go on and make a bunch of assets for the game, so probably won't get anything done when it comes to coding the game.