Scene loader, interaction highlights
I'm on the wire here. I really only have two options that I know of:
A) The Long Dark way, and that is to separate interiors and exteriors by making them separate scenes.
B) Teleport the player into buildings in the same scene.

A) would be preferable, as it would grant me more leeway when it comes to local lighting, AI and what have you.

I've been trying to get A working for the past two hours, but I'm not getting anywhere it.

Anyway, other changes (that I remember) are as follows:
-Added couple vertices to the zombie model's arm(s), meaning that Godot's awful inverse kinematics won't contort the zombie into what is essentially a spaghetti monster under some specific circumstances while blending animations together. Phew, that was a mouthful.

I also got some work on Industry done. Oh yeah, I also added a light tint when you're hovering your mouse over an object that's deemed to be interactable using.. the world's simplest shader. But I like it. It's paired up with a tween, so it blends in and out smoothly.

One last thing is that I added map choosing UI. It's bad. But it's functional. I'll come back to it, of course.

Okay, adieu!