Interaction UI, walking SFX, death animations and settings
Added interaction UI. One for when you hover over an object - that one gives you a prompt, like, F to open. The other gives you a response after interacting. For example, "This door's locked". Both are technically optional and have to be added on an item-by-item basis. Also added walking sound effects - they're not good in the slightest, but I'm no SFX designer. Also added death animations for zombies - there will be more variety eventually, but I'm likely going to take a step back and remodel zombies and come up with a clothing overlay system. That could save me time in the long run and make each zombie in the horde unique. Added a few settings option - fullscreen and resolution. I'm not satisfied with any of the UI, so I'm almost certainly going to come back and overhaul it.

I'll continue working on the map, as well as other assets. The Lakota textures are almost finished. I started working on a truck car model. Once I'm done with all that good stuff.. well.. we'll cross that bridge when we get there.