Final update.. for a while atleast
Admission exams are coming up in a little over two months and on top of that my dog passed away this Monday. I need to take a break. I'll come back. This isn't over.

Added some shaders and settings related to them for a slightly more retro look - colour dithering and internal resolution downscaling. Added a pistol viewmodel scene with a shooting animation. It's buggy, but it's a proof of concept.